LPS-63116 move to module
